"Messages client-server use JSON. Worked on mobile css.
Communication was reworked to use JSON instead of garbage string parsing. :)

Client-side CSS was updated for better responsiveness on mobile screens. App's height was unset and page dimensions were adjusted to utilize full viewport width and 90% of viewport height. The footer is now hidden on small screens for better viewing experience.

Also, a new 'whatnext.txt' file has been added to help track the future refactoring plans, debugging needs, and general to-do items for smooth project progression."
"Refactor client-server communication to use JSON

Refactored the communication between the client and the server to use JSON. The previous method was reliant on string parsing and was prone to errors. Switching to JSON communication increases consistency, readability, and reliability of the data being exchanged.